# app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAI Compatible API",
    "version": "1.0.0",
    "description": "OpenAI-compatible spec for LLM Django Server"
  },
  "servers": [
    {
      "url": "http://localhost:8000/v1"
    }
  ],
  "paths": {
    "/models": {
      "get": {
        "summary": "List available models",
        "operationId": "listModels",
        "responses": {
          "200": {
            "description": "A list of models",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelList"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/chat/completions": {
      "post": {
        "summary": "Create a chat completion",
        "operationId": "createChatCompletion",
        "requestBody": {
          "description": "Chat completion request",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_KEY"
      }
    },
    "schemas": {
      "ModelList": {
        "type": "object",
        "properties": {
          "object": {
            "type": "string",
            "example": "list"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "object": { "type": "string" },
                "owned_by": { "type": "string" }
              }
            }
          }
        }
      },
      "ChatMessage": {
        "type": "object",
        "properties": {
          "role": { "type": "string", "example": "user" },
          "content": { "type": "string", "example": "Hello" }
        }
      },
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["model", "messages"],
        "properties": {
          "model": { "type": "string" },
          "messages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ChatMessage" }
          },
          "max_tokens": { "type": "integer" }
        }
      },
      "ChatCompletionChoice": {
        "type": "object",
        "properties": {
          "index": { "type": "integer" },
          "message": {
            "type": "object",
            "properties": {
              "role": { "type": "string" },
              "content": { "type": "string" }
            }
          },
          "finish_reason": { "type": "string" }
        }
      },
      "ChatCompletionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "object": { "type": "string" },
          "model": { "type": "string" },
          "choices": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ChatCompletionChoice" }
          },
          "usage": {
            "type": "object",
            "properties": {
              "prompt_tokens": { "type": "integer" },
              "completion_tokens": { "type": "integer" },
              "total_tokens": { "type": "integer" }
            }
          }
        }
      }
    }
  }
}